requirements.txt
openai==0.26.4
pdf2image==1.16.2
Pillow==9.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
slate3k==0.5.3
